+++ src/controllers/index.ts
@@ -26,6 +26,13 @@ export const Controllers = createRegistry({
   RegimentEditor: () => import("@/controllers/regiment-editor").then(m => m.RegimentEditor),
   RegimentsOverview: () => import("@/controllers/regiments-overview").then(m => m.RegimentsOverview),
   ReligionsEditor: () => import("@/controllers/religions-editor").then(m => m.ReligionsEditor),
+  RiverCreator: () => import("@/controllers/river-creator").then(m => m.RiverCreator),
+  RiverEditor: () => import("@/controllers/river-editor").then(m => m.RiverEditor),
+  RiversOverview: () => import("@/controllers/rivers-overview").then(m => m.RiversOverview),
+  RouteCreator: () => import("@/controllers/route-creator").then(m => m.RouteCreator),
+  RouteEditor: () => import("@/controllers/route-editor").then(m => m.RouteEditor),
+  RouteGroupsEditor: () => import("@/controllers/route-groups-editor").then(m => m.RouteGroupsEditor),
+  RoutesOverview: () => import("@/controllers/routes-overview").then(m => m.RoutesOverview),
   StatesEditor: () => import("@/controllers/states-editor").then(m => m.StatesEditor),
   TradeAnimationEditor: () => import("@/controllers/trade-animation-editor").then(m => m.TradeAnimationEditor),
   TradeDetails: () => import("@/controllers/trade-details").then(m => m.TradeDetails),
