+++ src/index.html
@@ -8984,23 +8984,16 @@
     <script defer src="modules/ui/provinces-editor.js?v=1.131.0"></script>
     <script defer src="modules/ui/biomes-editor.js?v=1.112.0"></script>
     <script defer src="modules/ui/temperature-graph.js?v=1.106.6"></script>
-    <script defer src="modules/ui/routes-editor.js?v=1.131.0"></script>
-    <script defer src="modules/ui/routes-creator.js?v=1.120.5"></script>
-    <script defer src="modules/ui/route-group-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/ice-editor.js?v=1.111.0"></script>
     <script defer src="modules/ui/lakes-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/coastline-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/labels-editor.js?v=1.120.5"></script>
-    <script defer src="modules/ui/rivers-editor.js?v=1.131.0"></script>
-    <script defer src="modules/ui/rivers-creator.js?v=1.106.0"></script>
     <script defer src="modules/ui/relief-editor.js?v=1.99.00"></script>
     <script defer src="modules/ui/units-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/notes-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/ai-generator.js?v=1.120.5"></script>
     <script defer src="modules/ui/diplomacy-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/zones-editor.js?v=1.120.5"></script>
-    <script defer src="modules/ui/routes-overview.js?v=1.120.5"></script>
-    <script defer src="modules/ui/rivers-overview.js?v=1.120.5"></script>
     <script defer src="modules/ui/markers-overview.js?v=1.125.0"></script>
     <script defer src="modules/ui/emblems-editor.js?v=1.122.11"></script>
     <script defer src="modules/ui/markers-editor.js?v=1.122.11"></script>
